+++ backend/schemas/journal.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 class JournalBase(BaseModel):
     date: date
-    description: str
+    operation: str
     status: str
     amount: float
     entry_id: Optional[int] = None
@@ -13,8 +13,12 @@ class JournalCreate(JournalBase):
     """Схема для створення нового запису"""
     pass
 
+class JournalUpdate(JournalBase):
+    """Схема для оновлення існуючого запису"""
+    pass
+
 class JournalSchema(JournalBase):
     id: int
 
     class Config:
-        orm_mode = True
+        from_attributes = True